% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_DimPlot.R
\name{meb_DimPlot}
\alias{meb_DimPlot}
\title{meb_DimPlot}
\usage{
meb_DimPlot(
  seurat_obj,
  pt_size = 1.25,
  reduction = "umap",
  group_by = "orig.ident",
  cols = NULL,
  dims_plot = 1:2,
  split_by = NULL,
  label = F
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{pt_size}{point size - default 1.25}

\item{group_by}{metadata to group by - default "orig.ident"}

\item{cols}{color palette}

\item{label}{add labels to plot}
}
\value{
a ggplot object
}
\description{
Create Seurat DimPlot with consistent and attractive formatting
}
