% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_FeaturePlot.R
\name{meb_FeaturePlot}
\alias{meb_FeaturePlot}
\title{meb_FeaturePlot}
\usage{
meb_FeaturePlot(
  seurat_obj,
  features,
  pt_size = 1.25,
  reduction = "umap",
  scale_data = F,
  order = T
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{features}{genes to plot}

\item{pt_size}{point size - default 1.25}

\item{order}{}
}
\value{
a cowplot gtable with ggplot objects
}
\description{
Create Seurat FeaturePlot with consistent and attractive formatting
}
