% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_GSEAbarplot.R
\name{meb_GSEAbarplot}
\alias{meb_GSEAbarplot}
\title{meb_GSEAbarplot}
\usage{
meb_GSEAbarplot(
  result,
  n_terms,
  metric = "NES",
  fill,
  change_labels = TRUE,
  pval_fill = TRUE
)
}
\arguments{
\item{result}{output from fgsea analysis}

\item{n_terms}{number of terms to plot}

\item{metric}{which metric to plot - options are "NES" or "pval"}

\item{fill}{option to fill bar. if F, bars will all be black}

\item{change_labels}{reformat pathway/component labels - removes "GOCC" or "GOBP" header, removes dashes, changes to lowercase}

\item{pval_fill}{fills bars based on p value}
}
\value{
a ggplot object
}
\description{
Create consistently formatted bar plots from GSEA analysis output. Adapted from a similar function written by Miranda Hunter.
}
