% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_QC_plots.R
\name{meb_QC_plots}
\alias{meb_QC_plots}
\title{meb_QC_plots}
\usage{
meb_QC_plots(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\value{
grid of ggplot objects of QC metrics
}
\description{
Takes a Seurat object and produces plots for basic QC metrics: RNA counts, feature counts, % mito, log10 genes per UMI
}
