% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_rename_ensembleids.R
\name{meb_rename_ensemblids}
\alias{meb_rename_ensemblids}
\title{meb_rename_ensemblids}
\usage{
meb_rename_ensemblids(data)
}
\arguments{
\item{data}{a scRNAseq data matrix with counts and zebrafish ensembl ids}
}
\value{
a data matrix with ensembl ids translated into zebrafish gene names
}
\description{
Translate zebrafish ensembl ids in a matrix to gene names
}
