% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_translateEnsIDs_drerio.R
\name{meb_translateEnsIDs_drerio}
\alias{meb_translateEnsIDs_drerio}
\title{meb_translateEnsIDs_drerio}
\usage{
meb_translateEnsIDs_drerio(ensemblIDs = NULL)
}
\arguments{
\item{ensemblIDs}{a vector of danio rerio EnsemblIDs}
}
\value{
a matched vector danio rerio gene symbols
}
\description{
Translate a vector of zebrafish EnsemblIDs to gene symbols. Replaces un-mapped IDs with "un-translated".
}
