% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meb_translateEnsIDs_hsapiens.R
\name{meb_translateEnsIDs_hsapiens}
\alias{meb_translateEnsIDs_hsapiens}
\title{meb_translateEnsIDs_hsapiens}
\usage{
meb_translateEnsIDs_hsapiens(ensemblIDs = NULL)
}
\arguments{
\item{ensemblIDs}{a vector of human EnsemblIDs}
}
\value{
a matched vector human gene symbols
}
\description{
Translate a vector of human EnsemblIDs to gene symbols. Replaces un-mapped IDs with "un-translated". Can't translate EnsIDs with version numbers for now.
}
